requirements-dev.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-flask>=1.2.0
requests-mock>=1.11.0
coverage>=7.2.0
black>=23.7.0
flake8>=6.0.0
isort>=6.0.0
mypy>=1.5.0
types-requests>=2.31.0
types-PyYAML>=6.0.12